Refactor product  yaml fixture example
